Citizen_backend/requirements.txt
flask
flask-cors
psycopg2-binary
werkzeug
flask-jwt-extended
bcrypt
python-dotenv
